testing: Move testing logic to the providers package
The KubevirtProvider type is what provides the methods that run a node or run the k8s options.
Testing logic has been moved to a Base Provider Suite
