+++ cluster-provision/gocli/providers/base_provider_suite_test.go
@@ -1,0 +1,13 @@
+package providers
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestCmd(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Base Provider Suite")
+}
